% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{print hello}
\usage{
hello()
}
\description{
This does something less complicated than \code{\link[dplyr:mutate]{dplyr::mutate()}}
and \code{\link[dplyr:transmute]{dplyr::transmute()}}.
}
