% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R, R/doc.R
\docType{data}
\name{t3}
\alias{t3}
\alias{t4}
\alias{sharedDoc}
\title{Test data with NULL documented}
\format{An object of class \code{data.frame} with 10 rows and 2 columns.}
\usage{
t3

t4
}
\description{
Longer description for \code{t3} and \code{t4}
}
\keyword{datasets}
