% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{t1}
\alias{t1}
\alias{t2}
\title{Test data with shared description in this file}
\format{Something besides the default format}
\usage{
t1

t2
}
\description{
Longer description for \code{t1} and \code{t2}
}
\keyword{datasets}
